# eims-project/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: ExamFlow EIMS API
  description: |
    Exam Information Management System API for UP Tacloban.
    
    ## Authentication
    This API uses JWT Bearer tokens for authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```
    
    ## Rate Limiting
    - 100 requests per 15 minutes per IP
    - Rate limit headers included in responses
    
    ## WebSocket
    Real-time updates available at `ws://localhost:3001/ws?token=<access_token>`
  version: 1.0.0
  contact:
    name: ExamFlow Support
    email: support@examflow.up.edu.ph
  license:
    name: Educational Use
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api/v1
    description: Development server
  - url: https://examflow.up.edu.ph/api/v1
    description: Production server
...
paths:
...
  /results:
    get:
      tags: [Results]
      summary: List results
      description: Get results (filtered by role - students see only their own)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - name: examId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of results
    post:
      tags: [Results]
      summary: Create result
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateResultRequest'
      responses:
        '201':
          description: Result created
  /results/{id}/regrade:
    post:
      tags: [Results]
      summary: Request regrade
      description: Student requests a regrade for their result
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                  minLength: 20
                  maxLength: 2000
      responses:
        '201':
          description: Regrade request created
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
  schemas:
...
    CreateResultRequest:
      type: object
      required: [examId, studentId, score]
      properties:
        examId:
          type: string
          format: uuid
        studentId:
          type: string
          format: uuid
        score:
          type: number
          minimum: 0
        remarks:
          type: string
...